// prismicio-types.d.ts
type Simplify<T> = { [KeyType in keyof T]: T[KeyType] };
// ...
type HomePageDocumentDataSlicesSlice =
  | FooterSlice
  | ContactSectionSlice
  | ServicesSectionSlice
  | CoreValueSlice
  | HeroSectionSlice
  | CalltoActionSlice
  | BannerSlice;

/**
 * Content for Page d'accueil documents
 */
interface HomePageDocumentData {
  /**
   * Titre field in *Page d'accueil*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: home_page.titre_principal
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  titre_principal: prismic.KeyTextField;

  /**
   * Contenu Texte Principal field in *Page d'accueil*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: *None*
   * - **API ID Path**: home_page.contenu_texte
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/rich-text
   */
  contenu_texte: prismic.RichTextField;

  /**
   * Meta Title field in *Page d'accueil*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: home_page.meta_title
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  meta_title: prismic.KeyTextField;

  /**
   * Meta Description field in *Page d'accueil*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: home_page.meta_description
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  meta_description: prismic.KeyTextField;

  /**
   * Slice Zone field in *Page d'accueil*
   *
   * - **Field Type**: Slice Zone
   * - **Placeholder**: *None*
   * - **API ID Path**: home_page.slices[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/slices
   */
  slices: prismic.SliceZone<HomePageDocumentDataSlicesSlice>;
}

/**
 * Page d'accueil document from Prismic
 *
 * - **API ID**: `home_page`
 * - **Repeatable**: `false`
 * - **Documentation**: https://prismic.io/docs/content-modeling
 *
 * @typeParam Lang - Language API ID of the document.
 */
export type HomePageDocument<Lang extends string = string> =
  prismic.PrismicDocumentWithoutUID<
    Simplify<HomePageDocumentData>,
    "home_page",
    Lang
  >;